requirements.txt
openai
python-dotenv
langchain-community
pypdf
langchain
fastapi[standard]
werkzeug